requirements.txt
##################################################
# Medical Report Explanation System
# Python Dependencies
# Optimized for Mac M4 (Apple Silicon)
##################################################

# Core Framework
flask==3.0.0
flask-cors==4.0.0

# OpenAI Integration
openai>=1.12.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Environment Variables
python-dotenv==1.0.0

# Security & Validation
Werkzeug==3.0.1

# Optional: For enhanced features
# gunicorn==21.2.0  # Production WSGI server
# redis==5.0.1      # For caching (future enhancement)
requests>=2.31.0
